# specs/004-remote-photos-persistence/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Photo Admin Web API
  description: |
    REST API for photo collection management, analysis tool execution, and result visualization.

    **Key Features**:
    - Collection management (local/S3/GCS/SMB)
    - Configuration import/export (YAML ↔ Database)
    - Pipeline configuration (form-based editor)
    - Analysis tool execution (PhotoStats, Photo Pairing, Pipeline Validation)
    - Historical result queries and trend analysis

    **Authentication**: None (localhost deployment for v1)
  version: 1.0.0
  contact:
    name: Photo Admin Support
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
servers:
  - url: http://localhost:8000/api
    description: Local development server
...
paths:
...
  /collections:
    get:
      tags: [collections]
      summary: List all collections
      description: Returns all collections sorted by creation date (newest first)
      operationId: listCollections
      parameters:
        - name: state
          in: query
          description: Filter by collection state
          schema:
            type: string
            enum: [Live, Closed, Archived]
        - name: type
          in: query
          description: Filter by collection type
          schema:
            type: string
            enum: [local, s3, gcs, smb]
        - name: accessible_only
          in: query
          description: Show only accessible collections
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Collection list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collection'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [collections]
      summary: Create new collection
      description: Create a new collection with accessibility test
      operationId: createCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          description: Collection created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Collection name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /collections/{id}:
    get:
      tags: [collections]
      summary: Get collection by ID
      operationId: getCollection
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '200':
          description: Collection retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [collections]
      summary: Update collection
      description: Update collection name, location, state, or credentials
      operationId: updateCollection
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionUpdate'
      responses:
        '200':
          description: Collection updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Collection name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [collections]
      summary: Delete collection
      description: Delete collection with confirmation if results/jobs exist
      operationId: deleteCollection
      parameters:
        - $ref: '#/components/parameters/CollectionId'
        - name: force
          in: query
          description: Force deletion even if results/jobs exist
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Collection deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Collection has active jobs or results (use force=true)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Collection has 3 analysis results and 1 active job
                  result_count:
                    type: integer
                  active_job_count:
                    type: integer
                  requires_confirmation:
                    type: boolean
                    default: true
        '500':
          $ref: '#/components/responses/InternalServerError'
  /collections/{id}/test:
    post:
      tags: [collections]
      summary: Test collection accessibility
      description: Test if collection is accessible (filesystem path exists, remote credentials valid)
      operationId: testCollection
      parameters:
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '200':
          description: Accessibility test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessible:
                    type: boolean
                  error_message:
                    type: string
                    nullable: true
                  file_count:
                    type: integer
                    description: Number of files found (if accessible)
                    nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /collections/{id}/refresh:
    post:
      tags: [collections]
      summary: Refresh collection cache
      description: Manually invalidate file listing cache and fetch fresh listing
      operationId: refreshCollection
      parameters:
        - $ref: '#/components/parameters/CollectionId'
        - name: confirm
          in: query
          description: Confirm refresh if file count exceeds threshold
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Cache refreshed successfully or confirmation required
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      message:
                        type: string
                        example: Cache refreshed
                      file_count:
                        type: integer
                  - type: object
                    properties:
                      warning:
                        type: string
                        example: This collection has 150,000 files. Refreshing will incur API costs.
                      file_count:
                        type: integer
                      requires_confirmation:
                        type: boolean
                        default: true
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== CONFIGURATION ====================
...
components:
  parameters:
    CollectionId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Collection ID
...
  schemas:
...
    Collection:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [local, s3, gcs, smb]
        location:
          type: string
          description: Filesystem path or URI
        state:
          type: string
          enum: [Live, Closed, Archived]
        cache_ttl:
          type: integer
          nullable: true
          description: Cache TTL override (seconds)
        is_accessible:
          type: boolean
        last_error:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, type, location, state, is_accessible, created_at, updated_at]
    CollectionCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        type:
          type: string
          enum: [local, s3, gcs, smb]
        location:
          type: string
          minLength: 1
        state:
          type: string
          enum: [Live, Closed, Archived]
          default: Live
        cache_ttl:
          type: integer
          minimum: 1
          nullable: true
        credentials:
          type: object
          description: Remote storage credentials (type-specific)
          nullable: true
          oneOf:
            - $ref: '#/components/schemas/S3Credentials'
            - $ref: '#/components/schemas/GCSCredentials'
            - $ref: '#/components/schemas/SMBCredentials'
        metadata:
          type: object
          nullable: true
      required: [name, type, location]
    CollectionUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        location:
          type: string
          minLength: 1
        state:
          type: string
          enum: [Live, Closed, Archived]
        cache_ttl:
          type: integer
          minimum: 1
          nullable: true
        credentials:
          type: object
          nullable: true
        metadata:
          type: object
          nullable: true
    S3Credentials:
      type: object
      properties:
        aws_access_key_id:
          type: string
        aws_secret_access_key:
          type: string
        region:
          type: string
          default: us-east-1
      required: [aws_access_key_id, aws_secret_access_key]
    GCSCredentials:
      type: object
      properties:
        service_account_json:
          type: string
          description: GCS service account JSON (as string)
      required: [service_account_json]
    SMBCredentials:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        domain:
          type: string
          nullable: true
      required: [username, password]

    # Configuration schemas
...
    Error:
      type: object
      properties:
        message:
          type: string
        detail:
          type: string
          nullable: true
        errors:
          type: array
          items:
            type: object
          nullable: true
      required: [message]
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Validation error
            errors:
              - field: name
                message: Name is required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Collection not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: An unexpected error occurred
            detail: Database connection failed